Run eslint & refactor SpaceScene
